Unione di EnemyFiring e EnemyScript
I contenuti dei due script sono stati spostati nel singolo script EnemyScript.
