+++ Assets/Prefabs/Enemy.prefab
@@ -171,7 +171,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bcda6c2a92da8a64bb1222cf5389f19b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Bullet: {fileID: 893378567151498894, guid: ce4db71eaee18ee4fbceb821cc2ad153, type: 3}
 --- !u!114 &4973277965485738202
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -186,6 +185,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   spawner: {fileID: 0}
   deathSound: {fileID: -6141114204284630347}
+  Bullet: {fileID: 893378567151498894, guid: ce4db71eaee18ee4fbceb821cc2ad153, type: 3}
 --- !u!82 &-6141114204284630347
 AudioSource:
   m_ObjectHideFlags: 0
