--- Assets/Scripts/EnemyFiring.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bcda6c2a92da8a64bb1222cf5389f19b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
