+++ Assets/Scripts/EnemyScript.cs
@@ -7,11 +7,44 @@ public class EnemyScript : MonoBehaviour
 
     public AudioSource deathSound;
 
+    public GameObject Bullet;
+    private float timer = 0;
+    private float spawnRate = 2;
+
+
     // Start is called before the first frame update
     void Start()
     {
         spawner = GameObject.FindGameObjectWithTag("Respawn").GetComponent<EnemySpawner>();
         logic = GameObject.FindGameObjectWithTag("Logic").GetComponent<LogicScript>();
+    }
+
+    void Update()
+    {
+        if (timer < spawnRate)
+        {
+            timer += Time.deltaTime;
+        }
+        else
+        {
+            timer = 0;
+            FireBullet();
+        }
+    }
+
+    void FireBullet()
+    {
+        // preleviamo la posizione del player
+        GameObject playerShip = GameObject.Find("Ship");
+        if (playerShip != null)
+        {
+            GameObject bullet = Instantiate(Bullet);
+            bullet.transform.position = transform.position; // gli assegnamo la posizione dell'enemy
+                                                            // calcoliamo la direzione verso il player
+            Vector2 direction = playerShip.transform.position - bullet.transform.position;
+            // impostiamo la direzione del proiettile
+            bullet.GetComponent<EnemyBullet>().SetDirection(direction);
+        }
     }
 
     public void OnTriggerEnter2D(Collider2D other)
